Update mock data, update snapshot
