+++ apps/admin/src/pages/form/utils/filterFormResponse.spec.tsx
@@ -1,5 +1,5 @@
 import { filterFormResponse } from './filterFormResponse'
-import { mockData } from './filterFormResponseMocks'
+import { mockData } from './filterFormResponseMock'
 
 // This test compares the output of the function to a snapshot.
 // To update the snapshot, run the following command:
